IIIMS/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.32.0
cryptography==39.0.2
Django==4.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-jazzmin==2.6.0
django-reset-migrations==0.4.0
django-soft-delete==0.9.21
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.1.0
idna==3.4
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.1.0
mysqlclient==2.2.0
packaging==23.1
Pillow==9.4.0
protobuf==4.21.12
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==1.0.0
python-nginx==1.5.7
pytz==2022.7.1
pywin32==306
requests==2.29.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.15
